Remove parallelisation from IDPP. The overhead from creating new processes is larger than any gains from parallelisation.

Test for diels-alder with 200 IDPP images
 - MacOS - with parallelisation = 4.621 s, without parallelisation = 2.403 s
 - Windows - with parallelisation = 10.26 s, without parallelisation = 2.331 s

***

## Checklist

* [ ] The changes include an associated explanation of how/why
* [ ] Test pass
* [ ] Documentation has been updated
* [ ] Changelog has been updated