@t-young31 2023-09-28
I'm not convinced this gain is worth the additional code complexity. When a DFT calculation is _O_(minutes-hours) I'm not sure it makes sense to save _O_(seconds). Please convince me otherwise!