@shoubhikraj 2023-09-28
@t-young31 That is true, but now the IDPP is being used for i-EIP (and I am writing code for freezing-string method where IDPP is called every iteration). The problem is more obvious on windows (updated timings now) where processes are created, so the overhead is even higher. For POSIX, forking is cheaper so the overhead is less. I honestly don't think it increases the complexity of code that much. It just moves one part of the code outside parallel loop. What do you think.