@t-young31 2023-10-01
Agree – for testing speed and if called every iteration then it's worth the small amount more complexity